pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "kms"
version = "0.1.0"
description = "Local-first KMS control plane (scripts)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.1",
  "Jinja2>=3.1.0",
  "pypdf>=5.0.0",
  "Flask>=3.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0"]

[tool.setuptools.packages.find]
where = ["."]
include = ["kms*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
